grammar.cpp: Prefer SHIFT-ing strings over <Identifier>s
This commit fixes an error where the parser would prefer to consume <Identifier> over a specific keyword.

---- Explanation ----

What is "return"? Well, it's both the keyword AND an identifier (since it's not a number, a string literal, a char literal or punctuation).

When parsing a statement, we have an alternative that just has a single bare expression:
  expression -> Identifier
              | ...;
  statement  -> expression ";";
              | "return" expression ";"
              | ...;

Since an expression can be just an identifier (e.g. a variable name), the parser would incorrectly try to consume "return" as an Identifier. Whoops.